requirements.txt
langchain==0.2.17
langchain-community==0.2.17
langchain-huggingface
langchain-ollama
flask==3.0.3
sentence-transformers==3.2.1
pypdf==5.6.1
python-dotenv==1.0.1
pinecone-client==2.2.4
-e .